+++ front-end/src/app/layout/sidebar/menu-info.ts
@@ -1,4 +1,4 @@
-import { Report, ReportStatus } from 'app/shared/models';
+import { Report, ReportStatus } from 'app/shared/models/reports/report.model';
 import { MenuItem } from 'primeng/api';
 
 export enum ReportSidebarSection {
