+++ front-end/src/app/shared/models/reports/form-3.model.ts
@@ -8,8 +8,6 @@ import { TransactionTypes } from '../transaction.model';
 import { ScheduleATransactionTypes } from '../scha-transaction.model';
 import { ScheduleCTransactionTypes } from '../schc-transaction.model';
 import { ScheduleDTransactionTypes } from '../schd-transaction.model';
-import { ReportSidebarSection, MenuInfo } from 'app/layout/sidebar/menu-info';
-import { MenuItem } from 'primeng/api';
 
 export enum F3FormTypes {
   F3N = 'F3N',
@@ -166,19 +164,5 @@ export class Form3 extends BaseForm3 {
 
   static fromJSON(json: unknown): Form3 {
     return plainToInstance(Form3, json);
-  }
-
-  getMenuItems(sidebarSection: ReportSidebarSection, isEditable: boolean): MenuItem[] {
-    const transactionItems = [MenuInfo.manageTransactions(this), ...MenuInfo.addTransactions(this)];
-    return [
-      MenuInfo.enterTransaction(sidebarSection, isEditable, transactionItems),
-      MenuInfo.reviewTransactions(sidebarSection, this, isEditable),
-      MenuInfo.reviewReport(sidebarSection, [
-        ...MenuInfo.viewSummary(this),
-        MenuInfo.printPreview(this),
-        MenuInfo.addReportLevelMenu(this, isEditable),
-      ]),
-      MenuInfo.submitReport(sidebarSection, this, isEditable, 'SUBMIT YOUR REPORT'),
-    ];
   }
 }
